test: update type defs and convert refresh.test.js to .ts
